# [OPT] : add preprocess optimization for GLM4.5V and qwen2.5vl