… python/sglang/srt/multimodal/processors/qwen_vl.py