### Use Cases

- **AI Agents**  
  Multiple images are often passed as input, but usually only one is new. Caching prevents redundant preprocessing of repeated images.

- **Robotics / Navigation**  
  Frequent re-use of similar frames can benefit from fast cache retrieval and IPC-based transmission.

### How to Enable

```bash
export SGL_CACHE_MM_IMAGE=1
# set cache max tokenizer cache size,  or will use default value or max-available memorysize)
export SGL_TOKENIZER_CACHED_IMAGE_SIZE_MB=1024 #(max cache 1024MB image data)

```

### [TODO]
~~mange cache by using MB_SIZE instead of  image nums~~ (done)

### Notes

* Works with both qwen2-image-processor and glm4v-image-processor.

* CUDA IPC is automatically applied only in cache mode.

* Designed for scenarios where image reuse is common, improving both speed and efficiency.


### test_case and optimization:
python3 test_vlm_perf.py sglang 10 (cua steps)
qwen2.5vl-7B tp2 : 6914.15 ms -> 4998.39ms
GLM4.5V-108B tp4 : 7514.86ms ->  5472.50 ms


```python
from openai import OpenAI
import base64
import time
import sys
import os


PER_STEP_INPUT_IMAGE_NUM = 5 

server_type = sys.argv[1]

# set task step   example : 20,
task_step_num = int(sys.argv[2])

if server_type == "vllm":
    client = OpenAI(base_url=f"http://127.0.0.1:8000/v1", api_key="None")
elif server_type == "sglang":
    client = OpenAI(base_url=f"http://127.0.0.1:8080/v1", api_key="None")
else:
    print("not support mode")
   
# prepare some image data
warmup_image_path = "./test.png"


# prepare some image data in CUA folder (N + step -1)
# image should be named 0.png, 1.png......n.png 
img_path = "./test_noise"

paths = []
base64_strings = []
total_use_image_num = task_step_num + PER_STEP_INPUT_IMAGE_NUM -1


for idx in range(total_use_image_num):
    image_path = os.path.join(img_path, "{}.png".format(idx))
    paths.append(image_path)
    with open(image_path, "rb") as image_file:
        base64_strings.append(base64.b64encode(image_file.read()).decode("utf-8"))


with open(warmup_image_path, "rb") as image_file:
    encoded_string_warmup = base64.b64encode(image_file.read()).decode("utf-8")


model_path = "Qwen/Qwen2.5-VL-7B-Instruct"


show_time = lambda x, y : print("cost time {} ms".format((y-x) * 1000))

# warm up
response = client.chat.completions.create(
    model=model_path,
    messages=[
        {
            "role": "user",
            "content": [
                {
                    "type": "text",
                    "text": "please tell me what is in this image? try to describe in details",
                },
                {
                    "type": "image_url",
                    "image_url": {
                        "url": f"data:image/png;base64,{encoded_string_warmup}"
                    },
                },
            ],
        }
    ],
    max_tokens=32,
    temperature= 0,
)

print(response.choices[0].message.content)

#prepare  input data.

def generate_step_n_inputmessages(image_list, step_id, history_image_num = PER_STEP_INPUT_IMAGE_NUM):
    base_messages = CUA_messages = [
        {
            "role": "system",
            "content": "### 、、、\n# Role:资深在线旅游搜索策略专家\n\n## Profile:\n- Language: 中文\n- Description: 作为资深在线旅游搜索策略专家，具备深厚的行业经验和专业知识。熟悉携程在线旅游平台的规则与操作流程，尤其在酒店搜索、筛选及预订方面有着卓越的技能和敏锐的洞察力。能够精准分析用户需求，制定高效合理的搜索策略，以帮助用户快速找到符合需求的酒店。\n\n### Skill:\n1. 精通携程在线旅游平台的酒店搜索功能和页面布局。\n2. 擅长设定精准的搜索过滤条件，确保搜索结果的准确性。\n3. 具备强大的图像识别逻辑判断能力，能准确判断图片中是否包含特定元素。\n4. 熟悉酒店预订流程，可顺利完成下单操作。\n5. 能有条不紊地处理多步骤、循环式的搜索任务，确保流程不中断。\n\n## Goals:\n1. 在携程网站按照指定条件搜索酒店。\n2. 准确判断酒店房间图片中是否包含浴缸。\n3. 找到包含浴缸的酒店并完成预订下单。\n4. 若当前页面无符合要求酒店，能正确翻页继续搜索。\n5. 持续循环操作，直至成功找到带浴缸的酒店。\n\n## Constrains:\n1. 必须严格按照给定的位置、星级、设施等过滤条件进行搜索。\n2. 遇到位置访问弹窗，必须点击阻止。\n3. 按顺序依次处理酒店列表中的每一个酒店。\n4. 图片判断必须准确，不能误判。\n5. 预订下单流程要完整无误。\n\n## OutputFormat:\n1. 清晰记录搜索的每一步操作过程。\n2. 详细说明图片判断的依据和结果。\n3. 准确记录预订下单的酒店信息。\n4. 记录翻页操作的次数和原因。\n5. 最终输出成功找到带浴缸酒店的完整过程。\n\n## Workflow:\n1. First, 打开携程网站，进入酒店搜索页面，目的地部分先选中全部文字，然后填写**德国法兰克福**， 关键词部分先选中全部，然后输入“浴缸”\n2. Then,在搜索页面设置过滤条件，位置选择德国法兰克福，星级勾选五星，在高级筛选项中点击酒店设施并选择健身房，若弹出是否允许访问位置的弹窗，点击阻止。\n3. Next, 向下滑动整体页面高度的50%，查看酒店列表页面信息。\n4. 之后，按照酒店列表顺序，依次点击进入每一个酒店的详情页，接着点击酒店照片，再点击**房间类型**照片。\n5. 对于每张房间类型照片，仔细判断是否包含浴缸。若包含，则告知用户发现了带浴缸的酒店，然后进行预订下单操作；若不包含，则继续查看下一张房间图片，重复此步骤。\n6. 如果查询完所有图片都没有找到包含浴缸的图片，退出酒店详情页，回到酒店列表页面，选择下一个酒店，重复上述操作。\n7. 如果当前页面的酒店都不符合要求，点击下一页，继续从滑动页面开始重复上述所有步骤。\n8. Finally，持续循环操作，直到成功找到带浴缸的酒店\n9. 选择酒店第一个房型，点击左侧的房型照片部分，检查房型照片中是否包含浴缸\n10. 如果房型包含浴缸，则点击“预定”按钮进入到房型预定页面，在订房须读部分，进一步确认是否带有浴缸，如果有浴缸，则需在住客资源中，填写姓氏:zhang, 名字:san, 电子邮件:zhangsan@test.com,电话:16212112121, 填写完成后则滚动页面到最底部，每次滚动1000像素，要一直滚动到最底部，直到看到支付按钮\n10. 点击支付按钮，进行房间预定\n11. 进入到具体支付页面，就可以完成任务了，由用户接管选择具体的支付方式。\n\n## Initialization:\nAs a资深在线旅游搜索策略专家, you must follow the既定的搜索流程 and constraints, you must talk to user in default 中文，you must greet the user. Then introduce yourself and introduce the workflow.您好，我是资深在线旅游搜索策略专家，接下来我会按照您提供的要求，在携程网站上进行酒店搜索，具体流程是先设置过滤条件，然后滑动页面查看酒店列表，接着进入酒店详情页查看图片判断是否有浴缸，根据结果进行相应操作，不断循环直到找到带浴缸的酒店并完成预订。\n### 、、、"
        }
    ]
    
    start_idx, end_idx = 0, 0 
    if step_id < history_image_num:
        start_idx = 0
        end_idx = step_id + 1
    else:
        start_idx = step_id - history_image_num
        end_idx = start_idx + 1
    
    for idx in range(start_idx, end_idx):
        if idx > len(image_list)-1:
            print("out of data")
        
        insert_data_user =   {
            "role": "user",
            "content": [
                {
                    "type": "image_url",
                    "image_url": {
                        "url": f"data:image/any;base64,{image_list[idx]}"
                    }
                }
            ]
       }
        
        insert_data_agent =     {
            "role": "assistant",
            "content": "Action_Summary:" + "this is action{}_".format(step_id) +  "在页面中间偏上的“关键词（选填）”输入框中输入“浴缸”，该输入框位于“酒店级别”下拉框右侧，内部原本显示“机场/火车站/酒店名称...”的提示文字。\nAction: type(content=\"浴缸\")"
        }
        
        base_messages.append(insert_data_user)
        base_messages.append(insert_data_agent)
    
    return base_messages


all_messages = []

for idx in range(task_step_num):
    all_messages.append(generate_step_n_inputmessages(base64_strings, idx))


#start test
s_time = time.time()


while 1:
    try:
        message = all_messages.pop(0)
        response = client.chat.completions.create(
            model= model_path,
            messages = message,
            max_tokens=32,
            temperature= 0,
        )
    except Exception as e:
        print(e)
        print("task finish")
        break
e_time = time.time()
show_time(s_time, e_time)

```